Extract XML Reader logic for Styles into a separate XML Style Reader class
